cambiato tipo prezzo in BigDecimal

+++ src/org/java/pojo/Concerto.java
@@ -1,23 +1,24 @@
 package org.java.pojo;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalTime;
 
 public class Concerto extends Evento{
-	private double price;
+	private BigDecimal price;
 	private LocalTime hour;
 
-	public Concerto(String titolo, LocalDate data, int postiTotali, double price, LocalTime hour) throws Exception {
+	public Concerto(String titolo, LocalDate data, int postiTotali, BigDecimal price, LocalTime hour) throws Exception {
 		super(titolo, data, postiTotali);
 		setPrice(price);
 		setHour(hour);
 	}
 
-	public double getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
 
@@ -28,9 +29,7 @@ public class Concerto extends Evento{
 	public void setHour(LocalTime hour) {
 		this.hour = hour;
 	}
-	
-	
-	
+
 	@Override
 	public String toString() {
